datakit/requirements.txt
datetime==4.3
requests==2.25.1
mysql-connector-python==8.0.23
pandas==1.2.4
geopandas==0.9.0
numpy==1.20.1
geopy==2.1.0


